app,daemon rpc: now regtest example app against the harness
	Also decode JsonDetails for get_transactions and
	get_transaction_pool
