+++ old_rpc/get_transactions.go
@@ -8,55 +8,6 @@ import (
 	"fmt"
 )
 
-type OutPubKey string
-
-type EcdhInfo struct {
-	// Diffie-Helman elliptic curves' structures
-	Amount string `json:"amount"`
-	Mask   string `json:"mask"`
-}
-
-type RctSignatures struct {
-	// List of signatures used in the ring signatures truct - to hide the true origin
-	// of the transaction inputs.
-	Type   uint64      `json:"type"`
-	TxnFee uint64      `json:"txnFee"`
-	Ecdh   []EcdhInfo  `json:"ecdhInfo"`
-	OutPk  []OutPubKey `json:"outPk"`
-}
-
-type TaggedKey struct {
-	StealthKey string `json:"key"`
-	ViewTag    string `json:"view_tag"`
-}
-
-type Target struct {
-	TaggedKey `json:"tagged_key"`
-}
-type Output struct {
-	AmountCoinbase uint64 `json:"amount"`
-	Target         Target `json:"target"`
-}
-
-type PubKeyPrevOut struct {
-	Amount     uint64   `json:"amount"`
-	KeyOffsets []uint64 `json:"key_offsets"`
-	KeyImage   string   `json:"k_image"`
-}
-
-type Input struct {
-	Key PubKeyPrevOut `json:"key"`
-}
-
-type JsonDetails struct {
-	Version       uint64        `json:"version"`
-	TxUnlockTime  uint64        `json:"unlock_time"` // unlock time for txs we lock for n blocks
-	Vin           []Input       `json:"vin"`
-	Vout          []Output      `json:"vout"`
-	Extra         []uint8       `json:"extra"` // bytes: Payment ID, miner info, random bytes, etc.
-	RctSignatures RctSignatures `json:"rct_signatures"`
-}
-
 type TxEntry struct {
 	TxHash          string   `json:"tx_hash"`
 	BlockHeight     uint64   `json:"block_height"`
@@ -84,7 +35,7 @@ type GetTransactionsResponse struct {
 }
 
 type GetTransactionsRequest struct {
-	TxsHashes    []TxHash `json:"txs_hashes"`
+	TxsHashes    []string `json:"txs_hashes"`
 	DecodeAsJson bool     `json:"decode_as_json,omitempty"`
 }
